# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger documentation ApplicationAPI",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/customers/{customerID}/restore": {
            "post": {
                "tags": [
                    "Customer"
                ],
                "summary": "Restore a soft-deleted customer by ID",
                "description": "Restore a soft-deleted customer by ID.",
                "operationId": "restoreCustomerById",
                "parameters": [
                    {
                        "name": "customerID",
                        "in": "path",
                        "description": "ID of the soft-deleted customer to restore",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Customer restored successfully."
                    },
                    "204": {
                        "description": "No content"
                    },
                    "404": {
                        "description": "Customer not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Customer not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Ocorreu um erro ao processar a solicitação."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}